big_thing/tts_big_thing/requirements.txt
gtts
langdetect
python-vlc